Fibonacci container based solution.
- this solution reuses the fib db and fib counter modules to read and write to
the same dynamodb table.
- modified the code so the cached fibonacci sequence is pre populated in memory up to
50th element. To go any higher we'd have to use a big number library to prevent overflows.
- Added a k6 load test that ramps up to 1000 requests per second.
- Scaling is hard coded to maintain 5 running container instances at a time. This should more
easily support the ~1000 requests per second requirement.
- reduce the docker image size by only copying over required files (omitting node_modules, etc)

+++ functions/lambda/main.go
@@ -55,10 +55,10 @@ func Handler(request events.APIGatewayV2HTTPRequest) (events.APIGatewayProxyResp
 }
 
 type Response struct {
-	Value int `json:"value"`
+	Value uint64 `json:"value"`
 }
 
-func ok(value int) events.APIGatewayProxyResponse {
+func ok(value uint64) events.APIGatewayProxyResponse {
 	ret := &Response{
 		Value: value,
 	}
